[Bugfix] Update SerializedProperty extension method "GetParent" to support inherited private members. Move it into MySerializedProperty.cs
