+++ Extensions/EditorExtensions/MyGUI.cs
@@ -1,17 +1,15 @@
 ﻿#if UNITY_EDITOR
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Reflection;
 using UnityEngine;
 using UnityEditor;
 using Object = UnityEngine.Object;
 
 namespace MyBox.EditorTools
 {
-	public static class MyGUI
+    public static class MyGUI
 	{
 		#region Colors
 
@@ -315,66 +313,6 @@ namespace MyBox.EditorTools
 			}
 
 			GUI.enabled = guiState;
-		}
-
-		#endregion
-
-
-		#region SerializedProperty Get Parent
-
-		// Found here http://answers.unity.com/answers/425602/view.html
-		/// <summary>
-		/// Get parent object of SerializedProperty
-		/// </summary>
-		public static object GetParent(this SerializedProperty prop)
-		{
-			var path = prop.propertyPath.Replace(".Array.data[", "[");
-			object obj = prop.serializedObject.targetObject;
-			var elements = path.Split('.');
-			foreach (var element in elements.Take(elements.Length - 1))
-			{
-				if (element.Contains("["))
-				{
-					var elementName = element.Substring(0, element.IndexOf("[", StringComparison.Ordinal));
-					var index = Convert.ToInt32(element.Substring(element.IndexOf("[", StringComparison.Ordinal)).Replace("[", "").Replace("]", ""));
-					obj = GetValueAt(obj, elementName, index);
-				}
-				else
-				{
-					obj = GetValue(obj, element);
-				}
-			}
-
-			return obj;
-		}
-
-		private static object GetValue(object source, string name)
-		{
-			if (source == null)
-				return null;
-			var type = source.GetType();
-			var f = type.GetField(name, BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance);
-			if (f == null)
-			{
-				var p = type.GetProperty(name,
-					BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.IgnoreCase);
-				if (p == null)
-					return null;
-				return p.GetValue(source, null);
-			}
-
-			return f.GetValue(source);
-		}
-
-		private static object GetValueAt(object source, string name, int index)
-		{
-			var enumerable = GetValue(source, name) as IEnumerable;
-			if (enumerable == null) return null;
-
-			var enm = enumerable.GetEnumerator();
-			while (index-- >= 0)
-				enm.MoveNext();
-			return enm.Current;
 		}
 
 		#endregion
