+++ Extensions/EditorExtensions/MySerializedProperty.cs
@@ -1,5 +1,6 @@
 #if UNITY_EDITOR
 using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
@@ -38,7 +39,6 @@ namespace MyBox.EditorTools
 			for (int i = 0; i < property.arraySize; i++)
 				yield return property.GetArrayElementAtIndex(i);
 		}
-
 
 		/// <summary>
 		/// Replace array contents of SerializedProperty with another array 
@@ -119,6 +119,76 @@ namespace MyBox.EditorTools
 			if (fieldInfo == null) return false;
 			return Attribute.IsDefined(fieldInfo, typeof(T));
 		}
-	}
+
+        #region SerializedProperty Get Parent
+
+        // Found here http://answers.unity.com/answers/425602/view.html
+        // Update here https://gist.github.com/AdrienVR/1548a145c039d2fddf030ebc22f915de to support inherited private members.
+        /// <summary>
+        /// Get parent object of SerializedProperty
+        /// </summary>
+        public static object GetParent(this SerializedProperty prop)
+        {
+            var path = prop.propertyPath.Replace(".Array.data[", "[");
+            object obj = prop.serializedObject.targetObject;
+            var elements = path.Split('.');
+            foreach (var element in elements.Take(elements.Length - 1))
+            {
+                if (element.Contains("["))
+                {
+                    var elementName = element.Substring(0, element.IndexOf("[", StringComparison.Ordinal));
+                    var index = Convert.ToInt32(element.Substring(element.IndexOf("[", StringComparison.Ordinal)).Replace("[", "").Replace("]", ""));
+                    obj = GetValueAt(obj, elementName, index);
+                }
+                else
+                {
+                    obj = GetValue(obj, element);
+                }
+            }
+
+            return obj;
+        }
+
+        private static object GetValue(object source, string name)
+        {
+            if (source == null)
+                return null;
+
+            foreach (var type in GetHierarchyTypes(source.GetType()))
+            {
+                var f = type.GetField(name, BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance);
+                if (f != null)
+                    return f.GetValue(source);
+
+                var p = type.GetProperty(name, BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.IgnoreCase);
+                if (p != null)
+                    return p.GetValue(source, null);
+            }
+            return null;
+
+            IEnumerable<Type> GetHierarchyTypes(Type sourceType)
+            {
+                yield return sourceType;
+                while (sourceType.BaseType != null)
+                {
+                    yield return sourceType.BaseType;
+                    sourceType = sourceType.BaseType;
+                }
+            }
+        }
+
+        private static object GetValueAt(object source, string name, int index)
+        {
+            var enumerable = GetValue(source, name) as IEnumerable;
+            if (enumerable == null) return null;
+
+            var enm = enumerable.GetEnumerator();
+            while (index-- >= 0)
+                enm.MoveNext();
+            return enm.Current;
+        }
+
+        #endregion
+    }
 }
 #endif
